+++ src/content/exporters/exportersModule.ts
@@ -1,0 +1,10 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+/*
+ * This module contains code that should be loaded on-demand as ther is no reason
+ * to load it for every user out of the gate.
+ */
+
+export { exportAsHtml } from "./html";